+++ src/main/java/com/dlut/tradesys/service/SpecService.java
@@ -3,9 +3,9 @@ package com.dlut.tradesys.service;
 import com.dlut.tradesys.common.pojo.result.Result;
 
 public interface SpecService {
-    Result getSpec(Integer itemId);
+    Result getSpec(Long itemId);
 
-    Result addSpec(Integer itemId, String name);
+    Result addSpec(Long itemId, String name);
 
-    Result deleteSpec(Integer specId);
+    Result deleteSpec(Long specId);
 }
